+++ api/src/request/index.ts
@@ -1,6 +1,5 @@
 import type { AxiosInstance } from 'axios';
-import { requestInterceptor } from './request-interceptor.js';
-import { responseInterceptor } from './response-interceptor.js';
+import { ValidatingHttpAgent, ValidatingHttpsAgent } from './agent-with-ip-validation.js';
 
 export const _cache: { axiosInstance: AxiosInstance | null } = {
 	axiosInstance: null,
@@ -9,9 +8,11 @@ export const _cache: { axiosInstance: AxiosInstance | null } = {
 export async function getAxios() {
 	if (!_cache.axiosInstance) {
 		const axios = (await import('axios')).default;
-		_cache.axiosInstance = axios.create();
-		_cache.axiosInstance.interceptors.request.use(requestInterceptor);
-		_cache.axiosInstance.interceptors.response.use(responseInterceptor);
+
+		_cache.axiosInstance = axios.create({
+			httpAgent: new ValidatingHttpAgent(),
+			httpsAgent: new ValidatingHttpsAgent(),
+		});
 	}
 
 	return _cache.axiosInstance;